% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS后数据整理.R
\name{data_clean_for_analysis}
\alias{data_clean_for_analysis}
\title{Title}
\usage{
data_clean_for_analysis(data_for_plot, sample_path = "./sample_info.xlsx")
}
\arguments{
\item{sample_path}{}
}
\value{

}
\description{
Title
}
