% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NGS后数据整理.R
\name{find_nearest_sample}
\alias{find_nearest_sample}
\title{Title}
\usage{
find_nearest_sample(control_data, strain, first_nc_index, sample_ratio, y)
}
\arguments{
\item{y}{}
}
\value{

}
\description{
Title
}
