% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_analysis.R
\name{get_unique_num_table}
\alias{get_unique_num_table}
\title{Title}
\usage{
get_unique_num_table(
  data_for_analysis,
  mouse_index_list,
  strain,
  search_table = NULL
)
}
\arguments{
\item{strain}{}
}
\value{

}
\description{
Title
}
