% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{line_with_control}
\alias{line_with_control}
\title{Title}
\usage{
line_with_control(
  data,
  unique_n,
  control_unique_n = 0,
  y = "percent",
  show_name = T,
  grna_index = c(1:1000),
  remove_order = c(),
  limit = 0.5,
  seq = 0.1
)
}
\arguments{
\item{seq}{}
}
\value{

}
\description{
Title
}
