% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{marc_distribution_comparative}
\alias{marc_distribution_comparative}
\title{Title}
\usage{
marc_distribution_comparative(
  data_for_analysis,
  plot_index,
  y = "percent",
  group = "sample",
  logy = F,
  order = F,
  order_index = NULL,
  grna_index = 2:100
)
}
\arguments{
\item{grna_index}{}
}
\value{

}
\description{
Title
}
