% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{marc_distribution_inturn}
\alias{marc_distribution_inturn}
\title{Title}
\usage{
marc_distribution_inturn(
  data_for_analysis,
  plot_index,
  grna_index = 2:100,
  smooth = F,
  highlight = NULL
)
}
\arguments{
\item{smooth}{}
}
\value{

}
\description{
Title
}
