% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualization.R
\name{marc_logfc_inturn}
\alias{marc_logfc_inturn}
\title{Title}
\usage{
marc_logfc_inturn(
  data_for_analysis,
  plot_index,
  FC = "ratio_FC",
  grna_index = 2:100,
  logy = T
)
}
\arguments{
\item{logy}{}
}
\value{

}
\description{
Title
}
