% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean the ngs data.R
\name{normalization_with_beside}
\alias{normalization_with_beside}
\title{Title}
\usage{
normalization_with_beside(data, flank_num = 4)
}
\arguments{
\item{flank_num}{}
}
\value{

}
\description{
Title
}
