% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean the ngs data.R
\name{percent_marc}
\alias{percent_marc}
\title{Title}
\usage{
percent_marc(data_group, remove_grna_index = c(1, 3), remove_gene = "Cd47")
}
\arguments{
\item{remove_gene}{}
}
\value{

}
\description{
Title
}
